# feat: add alpha layer to colors and optimizations