## Edits
- Add alpha layer to ParseHexColor to support transparency.  #35 
- Removed the `fmt.Sscanf`'s to parse the colors manually, the functions is now a lot quicker (~120 times quicker).

## Benchmark
```go
package germanium

import "testing"

func BenchmarkParseHexColorSixChars(b *testing.B) {
	for i := 0; i < b.N; i++ {
		ParseHexColor("#112233")
	}
}

func BenchmarkParseHexColorThreeChars(b *testing.B) {
	for i := 0; i < b.N; i++ {
		ParseHexColor("#123")
	}
}

func BenchmarkParseHexColorNewSixChars(b *testing.B) {
	for i := 0; i < b.N; i++ {
		ParseHexColorNew("#112233")
	}
}

func BenchmarkParseHexColorNewThreeChars(b *testing.B) {
	for i := 0; i < b.N; i++ {
		ParseHexColorNew("#123")
	}
}
```

```go
$ go test -bench . -benchmem
goos: linux
goarch: amd64
pkg: github.com/matsuyoshi30/germanium
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
BenchmarkParseHexColorSixChars-4        	 2143605	       563.7 ns/op	      76 B/op	       6 allocs/op
BenchmarkParseHexColorThreeChars-4      	 2545946	       463.5 ns/op	      68 B/op	       3 allocs/op
BenchmarkParseHexColorNewSixChars-4     	311080213	         3.852 ns/op	       0 B/op	       0 allocs/op
BenchmarkParseHexColorNewThreeChars-4   	388633989	         3.100 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/matsuyoshi30/germanium	6.576s
```